backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pdfplumber==0.10.0
PyPDF2==3.0.1
python-dotenv==1.0.0